refactor: reorganize src/ into dashboard/ and mcp/ subfolders
Move all MCP backend files into src/mcp/:
- server.ts, router.ts, capabilities.ts, completion.ts, types.ts
- utils/, prompts/, resources/, storage/, tools/, tests/

Update all import paths throughout the codebase:
- src/dashboard/server.ts: ../storage -> ../mcp/storage etc.
- src/mcp/client.ts: ../capabilities -> ./capabilities etc.
- src/mcp/server.ts: ./mcp/session -> ./session etc.
- src/mcp/router.ts: ./mcp/session -> ./session etc.
- src/mcp/prompts/, resources/, tools/, tests/: fixed ../mcp/* -> ../* paths

Update build pipeline:
- package.json: dist/server.js -> dist/mcp/server.js
- bin/mcp-memory-server.js: updated dist path
- src/mcp/capabilities.ts: ../package.json -> ../../package.json

All 179 tests pass. TSC compiles with zero errors.
