+++ bin/mcp-memory-server.js
@@ -3,5 +3,5 @@ const sub = process.argv[2];
 if (sub === 'dashboard' || sub === 'mcp-memory-dashboard') {
   import('../dist/dashboard/server.js');
 } else {
-  import('../dist/server.js');
+  import('../dist/mcp/server.js');
 }
