+++ package.json
@@ -4,7 +4,7 @@
 	"description": "MCP Local Memory Service for coding copilot agents",
 	"mcpName": "io.github.vheins/local-memory-mcp",
 	"type": "module",
-	"main": "dist/server.js",
+	"main": "dist/mcp/server.js",
 	"bin": {
 		"local-memory-mcp": "bin/mcp-memory-server.js",
 		"mcp-memory-server": "bin/mcp-memory-server.js",
@@ -26,13 +26,13 @@
 	"author": "Muhammad Rheza Alfin <m.rheza.alfin@gmail.com>",
 	"license": "MIT",
 	"scripts": {
-		"build": "npm run dashboard:build && tsc && mkdir -p bin && printf \"#!/usr/bin/env node\\nimport '../dist/dashboard/server.js';\\n\" > bin/mcp-memory-dashboard.js && shx chmod +x dist/server.js dist/dashboard/server.js bin/mcp-memory-server.js bin/mcp-memory-dashboard.js",
+		"build": "npm run dashboard:build && tsc && mkdir -p bin && printf \"#!/usr/bin/env node\\nimport '../dist/dashboard/server.js';\\n\" > bin/mcp-memory-dashboard.js && shx chmod +x dist/mcp/server.js dist/dashboard/server.js bin/mcp-memory-server.js bin/mcp-memory-dashboard.js",
 		"dashboard:build": "vite build --config src/dashboard/ui/vite.config.ts",
 		"dashboard:dev": "vite dev --config src/dashboard/ui/vite.config.ts",
 		"prepare": "npm run build",
 		"dev": "tsc --watch",
-		"start": "node dist/server.js",
-		"start:mcp": "node dist/server.js",
+		"start": "node dist/mcp/server.js",
+		"start:mcp": "node dist/mcp/server.js",
 		"dashboard": "node dist/dashboard/server.js",
 		"both": "npm run start & npm run dashboard",
 		"test": "vitest --run",
