+++ src/mcp/storage/vectors.stub.ts
