+++ src/mcp/storage/vectors.ts
