+++ src/mcp/tests/dashboard.test.ts
