+++ src/mcp/tests/e2e.test.ts
