+++ src/mcp/tests/memory.search.test.ts
