+++ src/mcp/tests/normalize.test.ts
