+++ src/mcp/tests/query-expander.test.ts
