+++ src/mcp/tests/sqlite.test.ts
