+++ src/mcp/tests/v2-features.test.ts
