+++ src/mcp/tools/memory.acknowledge.ts
