+++ src/mcp/tools/memory.bulk-delete.ts
