+++ src/mcp/tools/memory.delete.ts
