+++ src/mcp/tools/memory.search.ts
