+++ src/mcp/tools/memory.summarize.ts
