+++ src/mcp/tools/task.bulk-manage.ts
