+++ src/mcp/types.ts
