+++ src/mcp/utils/completion.ts
