+++ src/mcp/utils/git-scope.ts
