+++ src/mcp/utils/mcp-response.ts
