+++ src/mcp/utils/normalize.ts
