+++ src/mcp/utils/pagination.ts
