+++ src/mcp/utils/query-expander.ts
